update default path pattern for StreamSupervisor (#639)
We sometimes see paths for the stream supervisor that look
like:

```
akka://test/user/StreamSupervisor-99961
```

This creates a lot of ids which aren't that useful. This
change updates the default pattern to just extract the
prefix and ignore the counter at the end. So activity for
all stream supervisors will show up as a single metric.